Pull request #3512: [BLUETOOTH-15424] MISRA fixes on common core 1 ble_port_fsci files
Merge in CONNINT/bluetooth_public from BLUETOOTH-15424-mcxw72-misra-check-on-ncp_loc_reader-and-loc_reader_host-applications to develop/main

Squashed commit of the following:

commit d29393f50934687c270037714d59e528a9b8600f
Author: Daniela Andreea Dumitrache <danielaandreea.dumitrache@nxp.com>
Date:   Mon May 12 17:41:52 2025 +0300

    [BLUETOOTH-15424] MISRA fixes on common core 1 ble_port_fsci files

+++ application/common/ble_port_fsci/core1/led/fsl_component_led.h
@@ -26,25 +26,25 @@
 #if (defined(gAppLedCnt_c) && (gAppLedCnt_c > 0))
 /* Start flashing all LEDs.*/
 #define LedStartFlashingAllLeds(void)         FSCI_transmitPayload(BLE_PORT_FSCI_OG,\
-                                                  g_LedStartFlashingAllLeds_c, NULL, 0, gFsciInterface_c);
+                                                  (uint8_t)g_LedStartFlashingAllLeds_c, NULL, 0, gFsciInterface_c);
 /* Stop flashing all LEDs */
 #define LedStopFlashingAllLeds(void)          FSCI_transmitPayload(BLE_PORT_FSCI_OG,\
-                                                  g_LedStopFlashingAllLeds_c, NULL, 0, gFsciInterface_c);
+                                                  (uint8_t)g_LedStopFlashingAllLeds_c, NULL, 0, gFsciInterface_c);
 /* Trun on LED1 */
 #define Led1On(void)                          FSCI_transmitPayload(BLE_PORT_FSCI_OG,\
-                                                  g_Led1On_c, NULL, 0, gFsciInterface_c);
+                                                  (uint8_t)g_Led1On_c, NULL, 0, gFsciInterface_c);
 /* Trun on LED2 */
 #define Led2On(void)                          FSCI_transmitPayload(BLE_PORT_FSCI_OG,\
-                                                  g_Led2On_c, NULL, 0, gFsciInterface_c);
+                                                  (uint8_t)g_Led2On_c, NULL, 0, gFsciInterface_c);
 /* Start flashing LED1 */
 #define  Led1Flashing(void)                   FSCI_transmitPayload(BLE_PORT_FSCI_OG,\
-                                                  g_Led1Flashing_c, NULL, 0, gFsciInterface_c);
+                                                  (uint8_t)g_Led1Flashing_c, NULL, 0, gFsciInterface_c);
 /* Start flashing LED2 */
 #define Led2Flashing(void)                    FSCI_transmitPayload(BLE_PORT_FSCI_OG,\
-                                                  g_Led2Flashing_c, NULL, 0, gFsciInterface_c);
+                                                  (uint8_t)g_Led2Flashing_c, NULL, 0, gFsciInterface_c);
 /* Trun off LED1 */
 #define Led1Off(void)                         FSCI_transmitPayload(BLE_PORT_FSCI_OG,\
-                                                  g_Led1Off_c, NULL, 0, gFsciInterface_c);
+                                                  (uint8_t)g_Led1Off_c, NULL, 0, gFsciInterface_c);
 #else
 /* Start flashing all LEDs.*/
 #define LedStartFlashingAllLeds(void)
